requirements.txt
pyqt5==5.15.2
opencv-python==4.11.0.86
torch==2.4.0
torchvision==0.19.0
numpy==1.26.4
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
